Backend/requirements.txt
SpeechRecognition
requests
pygame
gTTS
google-generativeai
librosa
numpy
pandas
scikit-learn
scipy
Flask
Flask-Cors 